refactor: 컴포넌트 이름은 대문자로
